requirements.txt
tuya-connector-python
ntplib==0.4.0
requests==2.31.0
python-dotenv